+++ src/type-info.c
@@ -98,6 +98,39 @@ enum vctrs_type vec_proxy_typeof(SEXP x) {
   return vec_base_typeof(x, true);
 }
 
+bool vec_is_list(SEXP x) {
+  if (TYPEOF(x) != VECSXP) {
+    return false;
+  }
+
+  switch(class_type(x)) {
+  // Bare list
+  case vctrs_class_none:
+    return true;
+
+  // Explicit lists
+  case vctrs_class_list:
+  case vctrs_class_list_of:
+    return true;
+
+  // Non-explicit S3 lists
+  case vctrs_class_unknown:
+    return vec_is_vector(x);
+
+  // TODO: Can this ever be considered a list?
+  case vctrs_class_rcrd:
+    return false;
+
+  // List-like classes known by `class_type()`.
+  // All data frame classes, posixlt.
+  default:
+    return false;
+  }
+}
+// [[ register() ]]
+SEXP vctrs_is_list(SEXP x) {
+  return Rf_ScalarLogical(vec_is_list(x));
+}
 
 // [[ include("vctrs.h") ]]
 bool vec_is_vector(SEXP x) {